[FEATURE styleguide] Allow arbitrary data to be associated with a tab
The motivation for this was to allow the Tabs component be used
to implement a smooth scrool navigation menu for the next version
of the edit-artwork form in CMS.

For that usage, it's helpful to attach a query selector to the Tab
so that we know what element we want to scroll to upon click.

This generalizes that idea to allow arbitrary data to be attached
to the Tab.