Prefer partner info to `collecting_institution` field in artwork sidebar
This commit updates the sidebar for the Responsive Artwork Page (RAP) to
prefer an artwork's associated partner information to the older
`collecting_institution` field. We would eventually like to deprecate
and remove the `collecting_institution` field altogether.

The `collecting_institution` field on an artwork is a string
representation of the institution loaning the work to another partner.

Quoting from a colleague here:

> Basically, at a number of different points we’ve wanted to clean up our
> institutional artwork records that predate the `Permanent Collection` /
> `On Loan` statuses, but it’s proved really tough because
> `collecting_institution` has historically been used for things that are
> _both_ on loan _and_ permanent collection.

Additional Context:

- https://artsy.slack.com/archives/CAZG82X4G/p1542669633601900
- https://artsy.slack.com/archives/CAZG82X4G/p1540477290000100
- https://github.com/artsy/gravity/pull/11790