Add event tracking for Read more / Learn more artist insight actions
When users are interacting with the Artist Insights 2.0 data, we want to
track user engagement (clicks) on the insights feature so that we can
understand if this information impacts commercial activity on Artsy.net.

The context submodule name is derived from the Insight's type. A couple
of these are created within the `SelectedCareerAchievements` component:

- `BLUE_CHIP`
- `HIGH_AUCTION`

The others are defined in Metaphysics:

- `SOLO_SHOW`
- `GROUP_SHOW`
- `REVIEWED`
- `COLLECTED`
- `BIENNIAL`

source: https://github.com/artsy/metaphysics/blob/master/src/schema/artist/insights.ts