Add gutter space between artist insight components
By setting the width of individual artist insight components set to 50%,
the text can occasionally brush up to the icons in the next columns.

Instead of adding padding, considering we don't know whether a
particiular insight will show up on the left or right side, limit the
Artist Insight component to 48% width and use `justify-content:
space-between`.