Prevent body from intercepting scroll events when modal is not open
Setting `overflow-y` to `auto` on body enables scrolling when the
content overflows. Since scroll events don't bubble on elements, they
get intercepted by body, and that causes window that typically listens
to scroll events to not receive them.

This issue was surfaced by a broken infinite scrolling experience. This
commit sets the `overflow-y` to `visible` which is the default.

https://drafts.csswg.org/cssom-view/#scrolling-events
https://drafts.csswg.org/cssom-view/#viewport