Bring back the cool ascii banner text in the README
This commit fixes the very cool ascii banner text in the README.

Wrapping the text in explicit code blocks should prevent it from
breaking again in the future. It's likely that a script ran some sort of
string strip command on the contents of the README, removing all leading
and trailing whitespace, causing the implicit code block, marked by
indentation, to break.

h/t to @orta for reporting :+1:

Addresses #2675