Loosen rules on phone numbers in AddressForm for international numbers.
Problem:

It was discovered that users weren't able to enter country codes or
white-space when providing a phone number when registering for an
auction. Moreover, due to the way that HTML5 validations work, users were
getting generic "Please match the requested format" messages, which is a
confusing UX.

Solution:

* Change the `pattern` attribute for the phone `Input` so that only
letters are not allowed in client side validations.
  + This is in-sufficient to ensure that all phone numbers are
    deliverable, but notably there are multiple ways to specify a phone
    number on Artsy, including the settings page on www.artsy.net, so
    additional constraints should likely be implemented at the API layer.
* Set the input type to `tel` to make the DOM as semantic and HTML5 nice
* Add a placeholder that includes a country code and will pass client
side validations.

Jira: https://artsyproduct.atlassian.net/browse/AUCT-687