WIP: Allow user to register+bid from ConfirmBid
NOTE: I think that there is a type bug that’s being worked around here.
MP assumes that there’s always a `Bidder` at `sale { registrationStatus }`
but this is incorrect. Make this query auth, but with a user that isn’t
registered for the (or any) sales: [query][mp_query]. Perhaps the type-
gen is totally working?


- commits WIP tracking test
- allow a user to register+bid if they have a credit card
- edit mutation response fixture fixtures to clarify new vs. existing
  bidder cases.

[mp_query]: https://metaphysics-staging.artsy.net/?query=query%20%7B%0A%20%20sale(id%3A%22training-video-mocktion%22)%20%7B%0A%20%20%20%20name%0A%20%20%20%20registrationStatus%20%7B%0A%20%20%20%20%20%20id%0A%20%20%20%20%20%20qualified_for_bidding%0A%20%20%20%20%7D%0A%20%20%7D%0A%7D