Fixes a test case where mockPostEvent was called twice when it should be called only once
The test above the test for analytics was polluting mock state since it
called `setTimeout` on the pending bidding status. Because of the delay,
the polling logic accidentally was getting kicked in even after the
actual test finished. As a result, the analytics call ended up being
called twice in the analtics test.