Allow users to request a condition report for an artwork in a sale
This commit introduces a "Request condition report" button on the
artwork page for artworks in an active sale.

The button will be displayed in the Condition show of the artwork page's
details component.

If the user is not currently logged in, we'll instead show a Log in
button, prompting the user to authenticate to request a condition
report.

After logging in to an existing account or creating a new account, the
user will be returned to the artwork page where the "Request condition
report" button will be available.

jira: https://artsyproduct.atlassian.net/browse/AUCT-743