Refactor live auction bid auction logic to clarify intent.
Clarify that we _only_ display “Watch Live Bidding” if it’s
_impossible_ for the user to become a qualifie bidder.

Some historical context is that when the artwork page was being
refactored to Reaction, there was a sense that it might take time for a
bidder to become qualified after attempting registration.

As such, the decision was made to only use the weaker “Watch” language,
instead of “Enter”, only if it was impossible for the user to bid.

I wonder how much of a concern it is now for there to be a delay in bidder
registration.

Notably, the current approach means that while a live sale is open for
bidding, and the bidder is on a blacklist, then we will _always_ display
“Enter” instead of “Watch”, when in reality, they will only be able to
watch unless an admin updates their bidder state.