Stop page from erroring when responding via conversations UI
There was a bug that caused the conversations UI to error out because the
timestamp on the message is null when optimistically supplied. I went back
and made the TimeSince component and all its helpers more robust to missing
data as well as ensured we only try to show TimeSince on a message if the
actual date is supplied.