Redirect to /log_in vs reloading upon initial 2FA setup (#3538)
* Sign out user via /user/sign_out before reauth redirect

Log out of the session and reload the page after initial 2FA enrollment.

* Use X-Requested-With HTTP header to tell artsy-passport to respond

This was annoying to figure out.

`DELETE /user/sign_out` hits this artsy-passport route:

https://github.com/artsy/artsy-passport/blob/cc01e392d5b46c2eac2dd7fc1c5c673740f24b7e/lib/app/logout.coffee#L16-L27

This route checks if `req.xhr`. The express documentation says that `req.xhr` is:

> A Boolean property that is true if the request’s X-Requested-With
> header field is “XMLHttpRequest”, indicating that the request was issued
> by a client library such as jQuery.

https://expressjs.com/en/api.html

So to get Force, via artsy-passport, to treat this request as an XHR request, we have to set this header.